# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty 100 Financial Intelligence Platform API",
    "description": "REST API server exposing financial analytics, stock screening, valuation, cashflow intelligence, PDF reports, and portfolio metrics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get 8-metric peer radar chart comparison data",
        "description": "Get 8-metric radar chart comparison data containing target company values, peer group average,\nand benchmark company values.\nReturns HTTP 404 if ticker is unknown.",
        "operationId": "get_company_peer_comparison_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerComparisonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get annual report documents and link validity for a company",
        "description": "Get list of annual report documents, BSE report URLs, and URL validity flags for a company.\nReturns HTTP 404 if no documents exist for the ticker.",
        "operationId": "get_company_documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDocumentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get company profile by ticker",
        "description": "Get detailed company profile and fundamental indicators by NSE ticker symbol.",
        "operationId": "get_company_profile_api_v1_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Company Profile Api V1 Companies  Ticker  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get all companies and latest KPIs in a sector",
        "description": "Get all companies belonging to the selected broad sector with their latest financial KPIs.\nReturns HTTP 404 if sector does not exist.",
        "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorCompaniesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/tearsheet/{ticker}": {
      "get": {
        "tags": [
          "Documents & Reports"
        ],
        "summary": "Download company tearsheet PDF",
        "description": "Download 2-page company tearsheet PDF by ticker.",
        "operationId": "get_tearsheet_pdf_api_v1_documents_tearsheet__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Tearsheet Pdf Api V1 Documents Tearsheet  Ticker  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{ticker}": {
      "get": {
        "tags": [
          "Documents & Reports"
        ],
        "summary": "Get annual report documents by ticker",
        "description": "Get annual report documents, report URLs, and URL validity flags for a company.\nReturns HTTP 404 if no documents exist for the ticker.",
        "operationId": "get_documents_by_ticker_api_v1_documents__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDocumentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompanyDocumentsResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "NSE Ticker Symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Company Name"
          },
          "total_documents": {
            "type": "integer",
            "title": "Total Documents",
            "description": "Total documents available"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "type": "array",
            "title": "Documents",
            "description": "List of company annual report documents"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name",
          "total_documents",
          "documents"
        ],
        "title": "CompanyDocumentsResponse"
      },
      "CompanyKPIs": {
        "properties": {
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe"
          },
          "roce": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roce"
          },
          "opm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opm"
          },
          "npm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npm"
          },
          "debt_to_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt To Equity"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio"
          },
          "pb_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Ratio"
          },
          "revenue_cagr_5yr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Cagr 5Yr"
          },
          "pat_cagr_5yr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pat Cagr 5Yr"
          },
          "free_cash_flow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Cash Flow"
          }
        },
        "type": "object",
        "title": "CompanyKPIs"
      },
      "DocumentItem": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year",
            "description": "Report Year"
          },
          "annual_report_title": {
            "type": "string",
            "title": "Annual Report Title",
            "description": "Annual Report Title"
          },
          "report_url": {
            "type": "string",
            "title": "Report Url",
            "description": "BSE/Company Report Download URL"
          },
          "is_url_valid": {
            "type": "boolean",
            "title": "Is Url Valid",
            "description": "Flag indicating if URL is available/valid"
          }
        },
        "type": "object",
        "required": [
          "annual_report_title",
          "report_url",
          "is_url_valid"
        ],
        "title": "DocumentItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeerComparisonResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "Target Company Ticker"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Target Company Name"
          },
          "peer_group_name": {
            "type": "string",
            "title": "Peer Group Name",
            "description": "Peer Group Name"
          },
          "benchmark_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Ticker",
            "description": "Benchmark Company Ticker"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metrics",
            "description": "8 Financial Metrics used for Radar Chart"
          },
          "company_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Company Values",
            "description": "Target company values/scores for 8 metrics"
          },
          "peer_group_average": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Peer Group Average",
            "description": "Peer group average values/scores"
          },
          "benchmark_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Benchmark Values",
            "description": "Benchmark company values/scores"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name",
          "peer_group_name",
          "metrics",
          "company_values",
          "peer_group_average",
          "benchmark_values"
        ],
        "title": "PeerComparisonResponse"
      },
...
      "SectorCompaniesResponse": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector",
            "description": "Selected Broad Sector Name"
          },
          "company_count": {
            "type": "integer",
            "title": "Company Count",
            "description": "Number of companies in sector"
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/SectorCompanyItem"
            },
            "type": "array",
            "title": "Companies",
            "description": "Companies in sector"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "company_count",
          "companies"
        ],
        "title": "SectorCompaniesResponse"
      },
      "SectorCompanyItem": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "NSE Ticker Symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Company Name"
          },
          "sub_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Sector",
            "description": "Sub-Sector Name"
          },
          "latest_kpis": {
            "$ref": "#/components/schemas/CompanyKPIs",
            "description": "Latest financial KPIs"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "company_name",
          "latest_kpis"
        ],
        "title": "SectorCompanyItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}